+++ src/tests/features/planning/usePlanCalculation.test.ts
@@ -4,13 +4,13 @@ import { createPinia, setActivePinia } from "pinia";
 import { flushPromises } from "@vue/test-utils";
 
 // Stores
-import { useGameDataStore } from "@/stores/gameDataStore";
 import { usePlanningStore } from "@/stores/planningStore";
 
 import {
 	materialsStore,
 	recipesStore,
 	buildingsStore,
+	exchangesStore,
 } from "@/database/stores";
 import { useMaterialData } from "@/database/services/useMaterialData";
 import { useBuildingData } from "@/database/services/useBuildingData";
@@ -39,21 +39,19 @@ vi.mock("@/database/services/usePlanetData", async () => {
 
 // Composables
 import { usePlanCalculation } from "@/features/planning/usePlanCalculation";
-import { until } from "@vueuse/core";
 
 describe("usePlanCalculation", async () => {
-	let gameDataStore: ReturnType<typeof useGameDataStore>;
 	let planningStore: ReturnType<typeof usePlanningStore>;
 
 	beforeAll(async () => {
 		setActivePinia(createPinia());
-		gameDataStore = useGameDataStore();
 		planningStore = usePlanningStore();
 
 		//@ts-expect-error mock data
 		await buildingsStore.setMany(buildings);
 		await recipesStore.setMany(recipes);
 		await materialsStore.setMany(materials);
+		await exchangesStore.setMany(exchanges);
 
 		const { preload } = useMaterialData();
 		const { preloadBuildings, preloadRecipes } = await useBuildingData();
@@ -62,10 +60,6 @@ describe("usePlanCalculation", async () => {
 		await preloadBuildings();
 		await preloadRecipes();
 		await flushPromises();
-
-		exchanges.forEach((e) => {
-			gameDataStore.exchanges[e.TickerId] = e;
-		});
 
 		vi.resetAllMocks();
 	});
@@ -91,7 +85,7 @@ describe("usePlanCalculation", async () => {
 	});
 
 	it("validate result", async () => {
-		const calculation = await usePlanCalculation(
+		const { calculate } = await usePlanCalculation(
 			// @ts-expect-error mock data
 			ref(plan_etherwind),
 			ref(undefined),
@@ -99,7 +93,7 @@ describe("usePlanCalculation", async () => {
 			ref(undefined)
 		);
 
-		const result = await until(calculation.result).toMatch((v) => v.done);
+		const result = await calculate();
 
 		expect(result.corphq).toBeFalsy();
 		expect(result.cogc).toBe("RESOURCE_EXTRACTION");
@@ -123,12 +117,12 @@ describe("usePlanCalculation", async () => {
 			ref(undefined),
 			ref(undefined)
 		);
-		await until(calculation.result.value.materialio.length).toMatch(
-			(v) => v !== 0
-		);
+		const result = await calculation.calculate();
 
-		const overviewData = await until(calculation.overviewData).toMatch(
-			(v) => v.roi !== Infinity
+		const overviewData = await calculation.calculateOverview(
+			result.materialio,
+			result.production,
+			result.infrastructure
 		);
 
 		expect(overviewData.dailyCost).toBe(38364.71626045736);
@@ -136,21 +130,17 @@ describe("usePlanCalculation", async () => {
 	});
 
 	it("validate visitationData", async () => {
-		const calculation = await usePlanCalculation(
+		const { calculate, visitationData } = await usePlanCalculation(
 			// @ts-expect-error mock data
 			ref(plan_etherwind),
 			ref(undefined),
 			ref(undefined),
 			ref(undefined)
 		);
-		await until(calculation.result.value.materialio.length).toMatch(
-			(v) => v !== 0
-		);
-		const visitationData = await until(calculation.visitationData).toMatch(
-			(v) => v.storageFilled !== Infinity
-		);
 
-		expect(visitationData.storageFilled).toBe(22.342256698594255);
+		const result = await calculate();
+
+		expect(visitationData.value.storageFilled).toBe(22.342256698594255);
 	});
 
 	it("validate existing and saveable", async () => {